fix(perf,test): perf-gate 回归修复——R3234-F/R3243-F/R3255-F 门禁与缓存归因 (R3254)
CI run 31630466921 全平台失败归因与修复：

- fix(browser,test): diff_ratio u8 加法溢出——CPU/GPU parity 差异大时 debug 构建
  panic（6 平台 "Run app tests" 失败主因，tests.rs:2788；改 u16 加法）
- fix(engine): ZW_SHAPED_LAYOUT 恢复显式 opt-in——R3234-F 默认开致 layout advance
  全量 shaping，perf-gate page/* layout_ms 37x 回归（welcome 1.7→60ms）
- fix(engine,layout,render): ZW_SHAPED_FALLBACK 恢复显式 opt-in——R3243-F 默认开
  致中文长文逐字多 face fallback shaping 每帧重排（morning paint 3x）；reftest
  实测关闭零回归（css-fonts 17/20 与全量均与默认一致）
- fix(render-foundation): face_match normal-stretch 快速路径——R3255-F 的
  available_widths 全表扫描 per-fragment，paint 4.5x 回归（medium 112→46ms）
- fix(render-foundation): shape_cache key 改字体字节 hash——instance_id 全局递增
  致每帧新 loader 全 miss（逐字长文每帧全量重排）；上限 4096→65536
- perf(render-foundation): fallback chain 预构建（逐 grapheme 去重分配消除）

验证：nextest app tests 410/410；clippy -D warnings clean；fmt clean；
reftest 全量 fallback 开/关对比无差异。

+++ apps/browser/src/tests.rs
@@ -2785,7 +2785,7 @@ fn local_composite_cpu_gpu_matrix_for_form_interactions() {
     let diff_ratio = |a: &[u8], b: &[u8]| -> f32 {
         let mut diff = 0usize;
         for (pa, pb) in a.chunks_exact(4).zip(b.chunks_exact(4)) {
-            let da = pa[0].abs_diff(pb[0]) + pa[1].abs_diff(pb[1]) + pa[2].abs_diff(pb[2]);
+            let da = pa[0].abs_diff(pb[0]) as u16 + pa[1].abs_diff(pb[1]) as u16 + pa[2].abs_diff(pb[2]) as u16;
             if da > 48 {
                 diff += 1;
             }
